Changed debt attribute of a person to a double to make calculations more precise

+++ commons/src/main/java/commons/Person.java
@@ -16,7 +16,7 @@ public class Person {
     protected String firstName;
     protected String lastName;
     protected String iban;
-    protected int debt;
+    protected double debt;
 
     //@JsonBackReference
     @ManyToOne
@@ -119,7 +119,7 @@ public class Person {
      * Setter for debt.
      * @param debt the email to set to
      */
-    public void setDebt(int debt) {
+    public void setDebt(double debt) {
         this.debt = debt;
     }
 
@@ -167,7 +167,7 @@ public class Person {
      * Returns the money that this person owes.
      * @return the ammount of money this person owes
      */
-    public int getDebt() {
+    public double getDebt() {
         return debt;
     }
 
@@ -222,6 +222,7 @@ public class Person {
      * @param add the amount to add to the dept
      */
     public void addDept(int add) {
-        debt+=add;
+        debt += add;
     }
+
 }
